Work with March Madness
The files in here are for a NCAA March Madness Bracket
currently, it is completely Random. It took a bit, but I tried using
more methods than I usually do. The main class is pretty straight
forward.
I need to work on better comments though. I understand what's going on,
But it would be better if others could understand my code as well.

All the best,
MJ :/