More File word
Given a wksh today to work on.
Not terribly difficult.
I changed the gibberish class to do all the giberish conversion in a
different, static method. Then, in InputGibberish, I just called that
static method to convert gibberish.
Need to work on writing to files in java. Reading easy.
P11 was pretty simple overall.
Changed the MeasureFile to run many times and total all lines, words,
and chars.
I think that's it.

All the best,
MJ :/