One, Two, Three ... ELEVENS!
Working on a lab given by college board. We've been working on the
basics, the classes Card and Deck as the base. We created a shuffling
method that wasn't too hard, something I've seen before. Elevens is a
solitaire game where you add things up to eleven. We got the Jar for it
to work, and I won first try.
Whoop whoop!!

All the best,
MJ :/