Big Update!!!
Working with a new lab!!
Pix Lab focusses on working with pictures and their pixels.
In the background, aside from it being fun and all,
It is supposed to be a good review for 2D arrays.
Overall, pretty easy.
However, it has a lot of interconnected classes.
We're learning about mirroring, altering colors, etc.
At the same time, we have out AP Baron's book that we're working on.
Other than that, nothing much going on.

All the best,
MJ :/