Linked List work
This has some wrap up for magpie stuff, but mostly a good portion
devoted to the start of Linked Lists. I like how it is set up overall,
and we are working soon on built in Linked Lists and then later linked
lists.

Linked Lists is a connected group of classes, where each node has the
reference to the next node. Fun stuff.

All the best,
MJ :/