Linked Lists
This unit featured linked lists and circular lists.
This was pretty fun, having each node of the list having a pointer to
the next node.
Overall, pretty good.
I liked the idea of the list, though it was a bit difficult to figure
out sorting the lists.
Anyways,
All the best,
MJ :