Queues and Stacks
I think Queue is a weird work. IDK.

Anyways, this is an alternative to Linked Lists, ish.
Queues involve added to the end and removing from the front.
	First on, first off.
In contrast, Stacks add to the top and remove from the top.
	First on, last off.

Anyways, it's pretty easy, though the priority Queue thing
doesn't make any sense really.

All the best,
MJ :/