Stacky Queues
Did some more programs with Stacks and Queues.
Movie line was using Queues.
Hardest was the P15_3, in which we had to basically make a calculator
that included order of operations. This was done via a stack and a
queue.
In order to make it function properly, I used recursion to get the
parenthesis evaluate it.
Otherwise, everything else should work dandy.

All the best,
MJ :/